+++ doc/source/conf.py
@@ -111,5 +111,7 @@ latex_documents = [
 # Example configuration for intersphinx: refer to the Python standard library.
 # intersphinx_mapping = {'http://docs.python.org/': None}
 
+suppress_warnings = ['ref.option']
+
 [extensions]
 # todo_include_todos = True